+++ app/lib/platform_config.rb
@@ -4,19 +4,12 @@ module PlatformConfig
   extend self
 
   def data(provider)
-    @data ||= {}
-    @data.fetch(provider) do |key|
-      @data[key] = load_data(provider)
-    end
+    Rails.cache.fetch(provider) { load_data(provider) }
   end
 
   private
 
   def load_data(provider)
-    JSON.parse(File.read(data_path(provider)))
-  end
-
-  def data_path(provider)
-    Rails.root.join("app/javascript/applications/CharKeeperApp/data/#{provider}.json")
+    JSON.parse(Rails.root.join("app/javascript/applications/CharKeeperApp/data/#{provider}.json").read)
   end
 end